requirements.txt
ccxt~=1.56.99
Django~=3.2.12
djangorestframework~=3.12.4
kucoin-futures-python
kucoin-python
psycopg2-binary
python-binance
pytz~=2021.3
websockets~=10.1
aiohttp~=3.8.1
requests~=2.26.0
Telethon~=1.24.0
numpy~=1.22.2
pandas~=1.4.1
ta~=0.9.0
celery~=5.2.3
redis~=4.2.1
django_tables2
django-filter
django-bootstrap-v5

pip~=20.0.2
rsa~=4.8
pyasn1~=0.4.8
amqp~=5.1.0
vine~=5.0.0
cryptography~=36.0.1
certifi~=2021.10.8
yarl~=1.6.3
idna~=3.3
multidict~=6.0.2
attrs~=21.4.0
kombu~=5.2.4
billiard~=3.6.4.0
setuptools~=59.6.0
cffi~=1.15.0
Deprecated~=1.2.13
packaging~=21.3
regex~=2022.1.18
pycares~=4.1.2
click~=8.1.1
sqlparse~=0.4.2
asgiref~=3.5.0
python-dateutil~=2.8.2
aiosignal~=1.2.0
frozenlist~=1.3.0
aiodns~=3.0.0
dateparser~=1.1.0
chardet~=4.0.0
tzlocal~=4.1
six~=1.16.0
wcwidth~=0.2.5
urllib3~=1.26.8
pyaes~=1.6.1
pyparsing~=3.0.7
pycparser~=2.21
wrapt~=1.14.0